refactor(cli): lift the flatpak runtime resolution
`gjsify ship --target flatpak` needs the same runtime/SDK triple, the same
finish-args defaults and the same `append-path` derivation that
`gjsify flatpak init` and `flatpak ci` already resolve, so they move out of
`commands/flatpak/` into `utils/flatpak-runtime.ts`.

Same move, same reason, as `utils/app-metadata.ts` when the AppStream and
desktop-entry renderers got a second caller: these are facts about the APP, not
about whichever command renders them, and the second copy is the one that
drifts. A `DEFAULT_GNOME_RUNTIME_VERSION` bumped in one place and not the other
would produce a `ship` Flatpak and a `flatpak init` Flatpak targeting different
GNOME releases from the same `package.json`.

`resolveRuntime` now takes a structural `RuntimeSelection` rather than a
`ConfigDataFlatpak`, because two config blocks answer that question from here on
and a resolver typed to one of them would have to be duplicated for the other.
`deriveAppendPath` was private in `init.ts` and is exported for the same reason.
